Demand read permissions once (#446)
* Demand read permissions once

If we can read a file we do a bunch of things like reading bytes to check if its an executable and inspecting it for a certificate.  Previously we called functions that Try to do each of those things, so if we get permission denied once we would keep trying other things triggering more exceptions.  Now we check if we can read once and pop out if we can't.

* Record SizeOnDisk

* Update dependencies.

* Move native methods from FileSystemCollector to NativeMethods.cs